Add a sword blink on right mouse, skill charges, and team spells tested on an ally dummy
Swords now carry Shadow Step as a fixed second weld: it steps behind body
after body, casting the primary key's skill at each one, untouchable until
the last blow, with its cooldown starting only then. Chain gems add steps.
Any skill can store charges through the new Stockpile support.

Team spells land on player characters through the ordinary hit queue:
Mending Word heals the ally you face, Battle Hymn empowers everyone around
you. Statuses now tick on characters too, and the arena has a wounded ally
dummy that is a real second player as far as the game is concerned.

+++ Assets/_Project/Code/Combat/Components/StatusPresentationComponents.cs
@@ -30,7 +30,7 @@ namespace TogetherWeFall.Combat
     public struct StatusVisual : IComponentData
     {
         /// <summary>Every status on this body, as a StatusMask. Zero is a clean one.</summary>
-        public ushort Icons;
+        public uint Icons;
 
         /// <summary>
         /// The colour to wash the body in, decided by the loudest status on it.
